♻️ parse correctly the formData
